requirements.txt
Flask
boto3
serverless_wsgi
werkzeug
python-dotenv